rule-logic-eval/pyproject.toml
[project]
name = "rule-logic-eval"
version = "0.1.0"
description = "Evaluation of experiments on rule logic benchamrk"
readme = "README.md"
authors = [
    { name = "Fabio Lorenzi", email = "fabio.lorenzi1@ibm.com" },
    { name = "Abi Langbridge", email = "abi.langbridge1@ibm.com" },
    { name = "Fearghal O'Donncha", email = "feardonn@ie.ibm.com" },
    { name = "Bradley Eck", email = "bradley.eck@ie.ibm.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "ott-jax>=0.5.0",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
    "typer>=0.16.0",
]

[project.scripts]
rulogic="rule_logic_eval.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
